refactor: 优化界面组件、日志框架及数据库迁移实现
- 重构数据库迁移脚本，提升健壮性与可维护性
- 将日志系统迁移至`tracing`框架，增强可观测性
- 统一界面数据刷新事件处理逻辑，实现状态集中管理
- 完善相关单元测试覆盖，确保核心功能可靠性

+++ crates/domain_manager/src/api/model/dns_operate.rs
@@ -6,16 +6,16 @@ use std::fmt::Display;
 #[derive(Serialize, Deserialize, Debug)]
 pub struct DnsOperateResponse {
     #[serde(rename = "TotalCount")]
-    total_count: i64,
+    total_count: i32,
 
     #[serde(rename = "PageSize")]
-    page_size: i64,
+    page_size: i32,
 
     #[serde(rename = "RequestId")]
     request_id: String,
 
     #[serde(rename = "PageNumber")]
-    page_number: i64,
+    page_number: i32,
 
     #[serde(rename = "RecordLogs")]
     pub record_logs: RecordLogs,
@@ -45,7 +45,7 @@ pub struct RecordLog {
     pub message: String,
 
     #[serde(rename = "ActionTimestamp")]
-    pub action_timestamp: i64,
+    pub action_timestamp: i32,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
