+++ crates/domain_manager/src/api/model/domain.rs
@@ -7,22 +7,22 @@ pub struct DomainQueryResponse {
     pre_page: bool,
 
     #[serde(rename = "CurrentPageNum")]
-    current_page_num: i64,
+    current_page_num: i32,
 
     #[serde(rename = "RequestId")]
     request_id: String,
 
     #[serde(rename = "PageSize")]
-    page_size: i64,
+    page_size: i32,
 
     #[serde(rename = "TotalPageNum")]
-    total_page_num: i64,
+    total_page_num: i32,
 
     #[serde(rename = "Data")]
     pub(crate) data: Data,
 
     #[serde(rename = "TotalItemNum")]
-    total_item_num: i64,
+    total_item_num: i32,
 
     #[serde(rename = "NextPage")]
     next_page: bool,
@@ -43,10 +43,10 @@ pub struct Domain {
     registration_date: String,
 
     #[serde(rename = "ExpirationCurrDateDiff")]
-    expiration_curr_date_diff: i64,
+    expiration_curr_date_diff: i32,
 
     #[serde(rename = "RegistrationDateLong")]
-    registration_date_long: i64,
+    registration_date_long: i32,
 
     #[serde(rename = "ResourceGroupId")]
     resource_group_id: String,
@@ -70,7 +70,7 @@ pub struct Domain {
     ccompany: String,
 
     #[serde(rename = "ExpirationDateLong")]
-    expiration_date_long: i64,
+    expiration_date_long: i32,
 
     #[serde(rename = "ExpirationDateStatus")]
     expiration_date_status: String,
